[BE] FIX: RoomController를 WebSocketController로 이름 변경 및 기능 수정 #174
